Thread Local Memory Mapping (TLMM).
This patch adds set of new system calls that permit user-level threads
to reserve part of the shared address space as a thread local memory
region.  Each thread in a process maps memory in its local region, so
that different physical pages might be mapped to the same virtual
address.  This patch works only on 64-bit x86 hardware.

Signed-off-by: Silas Boyd-Wickizer <sbw@mit.edu>